chore: move `BookFormContent`
