+++ feature/book/src/main/java/com/hisui/kanna/feature/book/NewBookScreen.kt
@@ -32,8 +32,6 @@ import com.hisui.kanna.core.model.Author
 import com.hisui.kanna.core.model.BookForm
 import com.hisui.kanna.core.model.BookReadStatus
 import com.hisui.kanna.core.model.BookStatus
-import com.hisui.kanna.feature.book.component.BookFormContent
-import com.hisui.kanna.feature.book.component.BookFormDialog
 import kotlinx.datetime.Clock
 
 @Composable
